split out fle options
